Changes in Business Layer for Run time get the Context Object

+++ MobileSchoolAPI/ParamModel/PARAMSTD.cs
@@ -10,7 +10,7 @@ namespace MobileSchoolAPI.Models
 
         public int STANDARDID { get; set; }
 
-        public Int64 USERID { get; set; }
+        public int USERID { get; set; }
         public string PASSWORD { get; set; }
     }
 }
